feat: add caching
