--- week2/cache.py
@@ -1,25 +1,0 @@
-from threading import Lock
-from typing import Any
-
-
-class Cache:
-    _instances = {}
-    _lock: Lock = Lock()
-
-    cache: dict
-
-    def __call__(cls, *args: Any, **kwargs: Any) -> Any:
-        with cls._lock:
-            if cls not in cls._instances:
-                instance = super().__call__(*args, **kwargs)
-                cls._instances[cls] = instance
-        return cls._instances[cls]
-
-    def __init__(self):
-        self.cache = {}
-
-    def get(self, key):
-        return self.cache.get(key)
-
-    def set(self, key, value):
-        self.cache[key] = value
