+++ week2/main.py
@@ -1,9 +1,9 @@
 import logging
 
-
 from fastapi import FastAPI, Request
 from fastapi.responses import JSONResponse
 
+from cache_response import cache_response
 from exception_handler import exception_handler
 from log_handler import log_request_handler
 from wiki import ProxyWiki
@@ -16,6 +16,7 @@ app = FastAPI()
 @app.get("/")
 @exception_handler
 @log_request_handler
+@cache_response
 def search_wiki(
     request: Request, query: str = None, message: str = None
 ) -> JSONResponse:
